fix canon lens type in exiv2
git-svn-id: file:///home/svn/slideshowgl/trunk/slideshowgl/slideshowgl-2.0.0@430 bbf4d130-548c-4f79-ac20-24e14d931773

+++ src/exif.c
@@ -121,7 +121,7 @@ enum exifrot {
 
 /* thread: load */
 enum rot imgexifgetrot(exifdata *exdat){
-	int rot;
+	long rot;
 #if HAVE_EXIV2
 	rot=exiv2getint(exdat,"Exif.Image.Orientation");
 #elif HAVE_EXIF
