+++ src/exiv2.h
@@ -13,7 +13,7 @@ struct exiv2data;
 struct exiv2data *exiv2init(const char *fn);
 void exiv2free(struct exiv2data *edata);
 char exiv2getstr(struct exiv2data *edata,const char **key,char *str,int len);
-int exiv2getint(struct exiv2data *edata,const char *key);
+long exiv2getint(struct exiv2data *edata,const char *key);
 
 #ifdef __cplusplus
 }
